Format sequence of dates

date_as_between_string() formats a sequence of dates to a single string that represents the sequence. The purpose is to simplify the request URL from including DATE=X&DATE=Y&DATE=Z to DATE = <>X,Z where X<Y<Z. The former is not only repetitive but invalid for the Megamation API.

Usage

date_as_between_string(.min, .max)

Arguments

.min

A minimum Date.

.max

A maximum Date.

Value

A single string formatted as "<>MM-DD-YYYY,MM-DD-YYYY" where the "<>" modifiers means "in-between".