megamation: Access 'DirectLine' Data Using the 'Megamation API'

Tools for accessing and handling 'DirectLine' data using the 'Megamation API' https://apidocs.megamation.com/. The API allows users to interact with data programmatically, instead of via the 'DirectLine' software. This package is community-maintained and is not officially supported by Megamation https://megamation.com/about-megamation/.

See also

Author

Maintainer: Adam Sadowski asebsadowski@gmail.com (ORCID)

Other contributors:

University of Guelph [copyright holder, funder]