Perform a Megamation API request

After creating a request with mm_req(), call mm_req_perform() to perform it with pagination and fetch the results back to R.

Usage

mm_req_perform(req, max_reqs = Inf)

Arguments

req

An API request generated by mm_req().

max_reqs

The maximum number of requests to perform. Use Inf to perform all requests until next_req() returns NULL.

Value

A list of HTTP responses. Each response is an S3 list with class httr2_response. (For more on this class, see httr2::response.) If the request was paginated, the responses correspond to pages.