Create a Megamation API request

mm_request() creates a request using httr2::request() and does the following:

Inserts the base URL using the environment variable MEGAMATION_URL from your .Renviron. Your key and base URL can be set using mm_set_creds().

Appends the URL with the endpoint defined by parameter endpoint.

Appends the URL with the filtering defined by parameter ... <dynamic-dots>.

Sets the user-agent as the GitHub megamation package.

Authenticates the request with HTTP basic authentication using environment variables MEGAMATION_KEY and MEGAMATION_URL from your .Renviron.

Handles HTTP errors so useful information from the response is extracted (e.g. "No response body").

Automatic caching of responses. See httr2::req_cache().

After creating a request with mm_request(), you can

Paginate the request using mm_req_paginate().

Perform the request and fetch the response using httr2::req_perform() or httr2::req_perform_iteratively() if the request is paginated.

Alternatively, mm_get() can at once define, paginate, and perform a request.

Usage

mm_request(endpoint, ..., allfields = TRUE, opts = req_opts())

Arguments

endpoint

The API endpoint. For example, "timecard" for employee transactions, and "workorder" for work orders. All endpoints are listed at https://apidocs.megamation.com/.

...

<dynamic-dots> Name-value pairs to filter the request. The name should be the lower-case name of a field that is filter-enabled (in Megamation's words, a criteria).

allfields

If TRUE, return all fields set to be available for the endpoint.

Value

An object of class httr2_request.